+++ pytsite/content/res/assets/js/content.js
@@ -1,8 +1,8 @@
 $(function () {
     var form = $('#odm-ui-form');
-    form.find('.widget-uid-title input').blur(function () {
-        var titleInput = $(this);
-        var descriptionInput = form.find('.widget-uid-description input');
+    form.find('.widget-uid-description input').focus(function () {
+        var descriptionInput = $(this);
+        var titleInput = form.find('.widget-uid-title input');
 
         if (descriptionInput.val() == '')
             descriptionInput.val(titleInput.val());